refactor: Merge frame layout helpers into dom/layout
Combine getFrameOffset and observeWindowsForLayoutShift, which both
position overlays over content in nested frames, into a single
dom/layout module. No behavior change.
